requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
Faker==36.1.1
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
